He will never - not good yet
I tried many many ideas. Does not work yet. My idea: convert number array to string array and then for loop and if statement will change the elements. 
I have a solution, to write manually every lines of the array and change the elements, but I believe there is a better way to do this. :)